+++ railties/lib/rails/generators/rails/authentication/authentication_generator.rb
@@ -1,8 +1,12 @@
 # frozen_string_literal: true
 
+require "rails/generators/bundle_helper"
+
 module Rails
   module Generators
     class AuthenticationGenerator < Base # :nodoc:
+      include BundleHelper
+
       class_option :api, type: :boolean,
         desc: "Generate API-only controllers and models, with no view templates"
 
@@ -27,7 +31,7 @@ module Rails
         template "app/views/passwords_mailer/reset.text.erb"
 
         template "test/mailers/previews/passwords_mailer_preview.rb"
-        template "test/helpers/session_test_helper.rb"
+        template "test/test_helpers/session_test_helper.rb"
       end
 
       def configure_application_controller
@@ -42,9 +46,9 @@ module Rails
       def enable_bcrypt
         if File.read(File.expand_path("Gemfile", destination_root)).include?('gem "bcrypt"')
           uncomment_lines "Gemfile", /gem "bcrypt"/
-          Bundler.with_original_env { execute_command :bundle, "install --quiet" }
+          bundle_command("install --quiet")
         else
-          Bundler.with_original_env { execute_command :bundle, "add bcrypt", capture: true }
+          bundle_command("add bcrypt", {}, quiet: true)
         end
       end
 
@@ -54,7 +58,7 @@ module Rails
       end
 
       def configure_test_helper
-        inject_into_file "test/test_helper.rb", "require_relative \"helpers/session_test_helper\"\n", after: "require \"rails/test_help\"\n"
+        inject_into_file "test/test_helper.rb", "require_relative \"test_helpers/session_test_helper\"\n", after: "require \"rails/test_help\"\n"
         inject_into_class "test/test_helper.rb", "TestCase", "    include SessionTestHelper\n"
       end