+++ railties/test/generators/authentication_generator_test.rb
@@ -75,7 +75,7 @@ class AuthenticationGeneratorTest < Rails::Generators::TestCase
 
     run_generator_instance
 
-    assert_includes @bundle_commands, [:bundle, "add bcrypt", { capture: true }]
+    assert_includes @bundle_commands, ["add bcrypt", {}, { quiet: true }]
   end
 
   def test_authentication_generator_with_api_flag
@@ -139,20 +139,18 @@ class AuthenticationGeneratorTest < Rails::Generators::TestCase
 
   private
     def run_generator_instance
-      commands = []
-      command_stub ||= -> (command, *args) { commands << [command, *args] }
+      @bundle_commands = []
+      command_stub ||= -> (command, *args) { @bundle_commands << [command, *args] }
 
       @rails_commands = []
       @rails_command_stub ||= -> (command, *_) { @rails_commands << command }
 
       content = nil
-      generator.stub(:execute_command, command_stub) do
+      generator.stub(:bundle_command, command_stub) do
         generator.stub(:rails_command, @rails_command_stub) do
           content = super
         end
       end
-
-      @bundle_commands = commands.filter { |command, _| command == :bundle }
 
       content
     end